Change back to loader.js
