+++ .eslintignore
@@ -1,0 +1,2 @@
+# ignore build file (loader.js)
+loader.js
