+++ .eslintrc.js
@@ -29,5 +29,5 @@ module.exports = {
     "globals": {
         "chrome": false,
         "googletag": false
-    }
+    },
 };
