+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "build": "npm run lint && webpack",
     "lint": "eslint ./src",
-    "start": "http-server ./dist/",
+    "start": "http-server ./src/",
     "watch": "webpack -w"
   },
   "keywords": [
