+++ src/bookmark-url.js
@@ -3,7 +3,7 @@
     var script = document.getElementById(id);
     if (!script) {
         script = document.createElement('script');
-        script.src = 'https://forbes.github.io/ad-inspector/dist/ad-inspector.bundle.js';
+        script.src = 'https://forbes.github.io/ad-inspector/src/loader.js';
         script.id = id;
         document.body.appendChild(script);
     }
