+++ webpack.config.js
@@ -3,14 +3,14 @@ var path = require('path');
 module.exports = {
     entry: './src/index.js',
     output: {
-        path: path.resolve(__dirname, 'dist'),
-        filename: 'ad-inspector.bundle.js'
+        path: path.resolve(__dirname, 'src'),
+        filename: 'loader.js'
     },
     module: {
         rules: [
             {
                 test: /\.js$/,
-                exclude: [/(node_modules)/, path.resolve(__dirname, 'src/background.js')],
+                exclude: [/(node_modules)/, path.resolve(__dirname, 'src/background.js'), path.resolve(__dirname, 'src/loader.js')],
                 use: {
                     loader: 'babel-loader',
                     options: {
